pyproject.toml
[tool.poetry]
name = "compdoc"
version = "0.1.0"
description = ""
authors = ["David Sillman <dsillman2000@gmail.com>"]
packages = [{ include = "compdoc" }]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.2"
docstring-parser = "^0.15"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
